requirements.txt
Django==1.6.5
django-allauth==0.17.0
django-model-utils==2.1.0
django-photologue==2.8.2
django-sortedm2m==0.7.0
gunicorn==19.2.1
oauthlib==0.6.3
Pillow==2.5.1
python-openid==2.2.5
pytz==2014.4
requests==2.3.0
requests-oauthlib==0.4.1
South==1.0
